Starting on the new calendar approach. Basics working.
The new approach: we're going to have a single calendar, tied to the
django site, global for everyone.  A user's office hours are the set
of occurrences of events he created.  The event's title is the
username, descriptions are blank. There are also provisions for basic
recurrences (courtesy of earlier work by ccg).